requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
arrow==1.3.0
    # via -r requirements.in
click==8.1.7
    # via
    #   click-default-group
    #   sqlite-utils
click-default-group==1.2.4
    # via sqlite-utils
colorama==0.4.6
    # via click
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
pluggy==1.5.0
    # via sqlite-utils
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   sqlite-utils
python-dotenv==1.0.1
    # via -r requirements.in
python-frontmatter==1.1.0
    # via -r requirements.in
pyyaml==6.0.2
    # via python-frontmatter
rich==13.7.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlite-fts4==1.0.3
    # via sqlite-utils
sqlite-utils==3.37
    # via -r requirements.in
tabulate==0.9.0
    # via sqlite-utils
types-python-dateutil==2.9.0.20240316
    # via arrow
